requirements.txt
asgiref==3.4.1
astroid==2.8.4
black==21.10b0
click==8.0.3
colorama==0.4.4
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
django-versatileimagefield==2.2
djangorestframework==3.12.4
djangorestframework-serializer-extensions==2.0.1
djangorestframework-simplejwt==5.0.0
drf-flex-fields==0.9.6
hashids==1.3.1
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
PyJWT==2.3.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-decouple==3.5
python-magic==0.4.24
python-magic-bin==0.4.14
pytz==2021.3
regex==2021.11.10
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
wrapt==1.13.3
